+++ lib/pow/store/backend/mnesia_cache.ex
@@ -6,9 +6,33 @@ defmodule Pow.Store.Backend.MnesiaCache do
   keys using the `expire` value. If the `expire` datetime is past, it'll
   send call the invalidator immediately.
 
+  ## Distribution
+
+  The MnesiaCache is built to handle multi-node setup.
+
+  If you initialize with `extra_db_nodes: Node.list()`, it'll automatically
+  connect to the cluster. If there is no other nodes available, the data
+  persisted to disk will be loaded, but if a cluster is running, the persisted
+  data (if any) will be purged on initialization, and the data from the cluster
+  will be loaded to the server.
+
+  When a cache key expires, the expiration will be verified before deletion to
+  ensure that it hasn't been updated by another node. When a key is updated on
+  a node, the node will ping all other nodes to refresh their invalidators so
+  the new TTL is used.
+
+  All nodes spun up will by default persist to disk. If you run multiple nodes
+  from the same directory you should make sure that each node has a separate
+  dir path configured. This can be done using different config files, or by
+  using system environment variables:
+
+      config :mnesia, dir: System.get_env("MNESIA_DIR")
+
   ## Initialization options
 
     * `:nodes` - list of nodes to use. This value defaults to `[node()]`.
+
+    * `:extra_db_nodes` - list of nodes in cluster to connect to.
 
     * `:table_opts` - options to add to table definition. This value defaults
       to `[disc_copies: nodes]`.
@@ -25,6 +49,8 @@ defmodule Pow.Store.Backend.MnesiaCache do
   """
   use GenServer
   alias Pow.{Config, Store.Base}
+
+  require Logger
 
   @behaviour Base
   @mnesia_cache_tab __MODULE__
@@ -63,7 +89,7 @@ defmodule Pow.Store.Backend.MnesiaCache do
   @impl GenServer
   @spec init(Config.t()) :: {:ok, map()}
   def init(config) do
-    table_init(config)
+    init_mnesia(config)
 
     {:ok, %{invalidators: init_invalidators(config)}}
   end
@@ -71,8 +97,10 @@ defmodule Pow.Store.Backend.MnesiaCache do
   @impl GenServer
   @spec handle_cast({:cache, Config.t(), binary(), any(), integer()}, map()) :: {:noreply, map()}
   def handle_cast({:cache, config, key, value, ttl}, %{invalidators: invalidators} = state) do
-    invalidators = update_invalidators(config, invalidators, key, ttl)
     table_update(config, key, value, ttl)
+
+    invalidators = update_invalidators(config, invalidators, key, ttl)
+    refresh_invalidators_in_cluster(config)
 
     {:noreply, %{state | invalidators: invalidators}}
   end
@@ -85,14 +113,35 @@ defmodule Pow.Store.Backend.MnesiaCache do
     {:noreply, %{state | invalidators: invalidators}}
   end
 
+  @spec handle_cast({:refresh_invalidators, Config.t()}, map()) :: {:noreply, map()}
+  def handle_cast({:refresh_invalidators, config}, %{invalidators: invalidators} = state) do
+    clear_invalidators(invalidators)
+
+    {:noreply, %{state | invalidators: init_invalidators(config)}}
+  end
+
   @impl GenServer
   @spec handle_info({:invalidate, Config.t(), binary()}, map()) :: {:noreply, map()}
   def handle_info({:invalidate, config, key}, %{invalidators: invalidators} = state) do
     invalidators = clear_invalidator(invalidators, key)
-
-    table_delete(config, key)
+    invalidators =
+      config
+      |> fetch(key)
+      |> delete_or_reschedule(config, invalidators)
 
     {:noreply, %{state | invalidators: invalidators}}
+  end
+
+  defp delete_or_reschedule(nil, _config, invalidators), do: invalidators
+  defp delete_or_reschedule({key, _value, key_config, expire}, config, invalidators) do
+    case Enum.max([expire - timestamp(), 0]) do
+      0 ->
+        table_delete(config, key)
+
+        invalidators
+      ttl ->
+        update_invalidators(key_config, invalidators, key, ttl)
+    end
   end
 
   defp update_invalidators(config, invalidators, key, ttl) do
@@ -100,6 +149,19 @@ defmodule Pow.Store.Backend.MnesiaCache do
     invalidator  = trigger_ttl(config, key, ttl)
 
     Map.put(invalidators, key, invalidator)
+  end
+
+  defp refresh_invalidators_in_cluster(config) do
+    :running_db_nodes
+    |> :mnesia.system_info()
+    |> Enum.reject(& &1 == node())
+    |> Enum.each(&:rpc.call(&1, GenServer, :cast, [__MODULE__, {:refresh_invalidators, config}]))
+  end
+
+  defp clear_invalidators(invalidators) do
+    Enum.reduce(invalidators, invalidators, fn {key, _ref}, invalidators ->
+      clear_invalidator(invalidators, key)
+    end)
   end
 
   defp clear_invalidator(invalidators, key) do
@@ -112,16 +174,22 @@ defmodule Pow.Store.Backend.MnesiaCache do
   end
 
   defp table_get(config, key) do
+    config
+    |> fetch(key)
+    |> case do
+      {_key, value, _config, _expire} -> value
+      nil                             -> :not_found
+    end
+  end
+
+  defp fetch(config, key) do
     mnesia_key = mnesia_key(config, key)
 
     {@mnesia_cache_tab, mnesia_key}
     |> :mnesia.dirty_read()
     |> case do
-      [{@mnesia_cache_tab, ^mnesia_key, {_key, value, _config, _expire}} | _rest] ->
-        value
-
-      [] ->
-        :not_found
+      [{@mnesia_cache_tab, ^mnesia_key, {_key, value, config, expire}} | _rest] -> {key, value, config, expire}
+      [] -> nil
     end
   end
 
@@ -130,7 +198,7 @@ defmodule Pow.Store.Backend.MnesiaCache do
     expire     = timestamp() + ttl
     value      = {key, value, config, expire}
 
-    :mnesia.transaction(fn ->
+    :mnesia.sync_transaction(fn ->
       :mnesia.write({@mnesia_cache_tab, mnesia_key, value})
     end)
   end
@@ -138,7 +206,7 @@ defmodule Pow.Store.Backend.MnesiaCache do
   defp table_delete(config, key) do
     mnesia_key = mnesia_key(config, key)
 
-    :mnesia.transaction(fn ->
+    :mnesia.sync_transaction(fn ->
       :mnesia.delete({@mnesia_cache_tab, mnesia_key})
     end)
   end
@@ -146,10 +214,17 @@ defmodule Pow.Store.Backend.MnesiaCache do
   defp table_keys(config, opts \\ []) do
     namespace = mnesia_key(config, "")
 
-    @mnesia_cache_tab
-    |> :mnesia.dirty_all_keys()
+    sync_all_keys()
     |> Enum.filter(&String.starts_with?(&1, namespace))
     |> maybe_remove_namespace(namespace, opts)
+  end
+
+  defp sync_all_keys() do
+    {:atomic, keys} = :mnesia.sync_transaction(fn ->
+      :mnesia.all_keys(@mnesia_cache_tab)
+    end)
+
+    keys
   end
 
   defp maybe_remove_namespace(keys, namespace, opts) do
@@ -163,25 +238,154 @@ defmodule Pow.Store.Backend.MnesiaCache do
     end
   end
 
-  defp table_init(config) do
-    nodes      = Config.get(config, :nodes, [node()])
-    table_opts = Config.get(config, :table_opts, disc_copies: nodes)
-    table_def  = Keyword.merge(table_opts, [type: :set])
-    timeout    = Config.get(config, :timeout, :timer.seconds(15))
-
-    case :mnesia.create_schema(nodes) do
-      :ok                                 -> :ok
-      {:error, {_, {:already_exists, _}}} -> :ok
+  defp init_mnesia(config) do
+    config
+    |> find_active_cluster_nodes()
+    |> case do
+      []    -> init_cluster(config)
+      nodes -> join_cluster(config, nodes)
     end
+  end
 
-    :rpc.multicall(nodes, :mnesia, :start, [])
+  defp find_active_cluster_nodes(config) do
+    visible_nodes = Node.list()
+    db_nodes      = Config.get(config, :extra_db_nodes, [])
+
+    db_nodes
+    |> Enum.filter(& &1 in visible_nodes)
+    |> Enum.filter(&:rpc.block_call(&1, :mnesia, :system_info, [:is_running]) == :yes)
+  end
+
+  defp init_cluster(config) do
+    nodes = nodes(config)
+
+    with :ok <- start_mnesia(nodes),
+         :ok <- change_table_copy_type(config, nodes),
+         :ok <- create_table(config, nodes),
+         :ok <- wait_for_table(config) do
+
+      Logger.info("[#{inspect __MODULE__}] Mnesia cluster initiated on #{inspect nodes}")
+    else
+      {:error, reason} ->
+        Logger.error("[inspect __MODULE__}] Couldn't initialize mnesia cluster because: #{inspect reason}")
+        {:error, reason}
+    end
+  end
+
+  defp join_cluster(config, cluster_nodes) do
+    nodes = nodes(config)
+
+    with :ok <- reset_mnesia(nodes, cluster_nodes),
+         :ok <- connect_to_cluster(nodes, cluster_nodes),
+         :ok <- change_table_copy_type(config, nodes),
+         :ok <- sync_table(config, nodes, cluster_nodes),
+         :ok <- wait_for_table(config) do
+
+      Logger.info("[#{inspect __MODULE__}] Joined mnesia cluster nodes #{inspect cluster_nodes} for #{inspect nodes}")
+
+      :ok
+    else
+      {:error, reason} ->
+        Logger.error("[inspect __MODULE__}] Couldn't join mnesia cluster because: #{inspect reason}")
+        {:error, reason}
+    end
+  end
+
+  defp nodes(config), do: Config.get(config, :nodes, [node()])
+
+  defp start_mnesia(nodes) when is_list(nodes) do
+    Enum.each(nodes, &:ok = start_mnesia(&1))
+  end
+  defp start_mnesia(node) do
+    node
+    |> block_call(Application, :start, [:mnesia])
+    |> case do
+      {:error, {:already_started, :mnesia}} -> :ok
+      :ok                                   -> :ok
+    end
+  end
+
+  defp reset_mnesia(nodes, cluster_nodes) when is_list(nodes) do
+    Enum.each(nodes, &:ok = reset_mnesia(&1, cluster_nodes))
+  end
+  defp reset_mnesia(node, [cluster_node | _rest]) do
+    block_call(node, Application, :stop, [:mnesia])
+
+    :ok = block_call(node, :mnesia, :delete_schema, [[node]])
+    {:atomic, :ok} = block_call(cluster_node, :mnesia, :del_table_copy, [:schema, node])
+
+    start_mnesia(node)
+
+    :ok
+  end
+
+  defp change_table_copy_type(config, nodes) when is_list(nodes) do
+    Enum.each(nodes, &:ok = change_table_copy_type(config, &1))
+  end
+  defp change_table_copy_type(config, node) do
+    copy_type = get_copy_type(config, node)
+
+    case :mnesia.change_table_copy_type(:schema, node, copy_type) do
+      {:atomic, :ok}                               -> :ok
+      {:aborted, {:already_exists, :schema, _, _}} -> :ok
+    end
+  end
+
+  defp get_copy_type(config, node) do
+    types      = [:ram_copies, :disc_copies, :disc_only_copies]
+    table_opts = Config.get(config, :table_opts, [])
+
+    Enum.find(types, :disc_copies, fn type ->
+      nodes = table_opts[type] || []
+
+      node in nodes
+    end)
+  end
+
+  defp create_table(config, nodes) do
+    table_opts = Config.get(config, :table_opts, [disc_copies: nodes])
+    table_def  = Keyword.merge(table_opts, [type: :set])
 
     case :mnesia.create_table(@mnesia_cache_tab, table_def) do
       {:atomic, :ok}                                   -> :ok
       {:aborted, {:already_exists, @mnesia_cache_tab}} -> :ok
     end
+  end
 
-    :ok = :mnesia.wait_for_tables([@mnesia_cache_tab], timeout)
+  defp sync_table(config, nodes, cluster_nodes) when is_list(nodes) do
+    Enum.each(nodes, &{:atomic, :ok} = sync_table(config, &1, cluster_nodes))
+  end
+  defp sync_table(_config, node, [cluster_node | _rest]) do
+    copy_type = block_call(cluster_node, :mnesia, :table_info, [@mnesia_cache_tab, :storage_type])
+
+    block_call(node, :mnesia, :add_table_copy, [@mnesia_cache_tab, node, copy_type])
+  end
+
+  defp block_call(node, module, method, args) do
+    this_node = node()
+
+    case node do
+      ^this_node -> apply(module, method, args)
+      node       -> :rpc.block_call(node, module, method, args)
+    end
+  end
+
+  defp wait_for_table(config) do
+    timeout = Config.get(config, :timeout, :timer.seconds(15))
+
+    :mnesia.wait_for_tables([@mnesia_cache_tab], timeout)
+  end
+
+  defp connect_to_cluster(nodes, cluster_nodes) when is_list(nodes) do
+    Enum.each(nodes, &:ok = connect_to_cluster(&1, cluster_nodes))
+  end
+  defp connect_to_cluster(node, [cluster_node | _cluster_nodes]) do
+    node
+    |> block_call(:mnesia, :change_config, [:extra_db_nodes, [cluster_node]])
+    |> case do
+      {:ok, _}         -> :ok
+      {:error, reason} -> {:error, reason}
+    end
   end
 
   defp mnesia_key(config, key) do